moe_fp8_boundary_qag redesign: per-channel[1,1,mlp] dynamic + STE, wo bf16, no e4m3 grad cast
- _q_boundary: DYNAMIC per-mlp-channel scale [1,1,mlp] (shared across experts), stop_gradient (STE).
  w0/w1 only; wo STAYS bf16 (per-out-channel over its embed gather axis is unsound).
- _ste_quant: STE custom_vjp (e4m3 fwd, bf16 d w=g/sc bwd) -- weight grad never e4m3.
- ops._gmm_bwd: drhs QArray cotangent keeps NATURAL bf16/f32 dtype (was astype e4m3 = the NaN +
  reduce-scatter blowup). Per-channel zero scale cotangent.
- wo QArray build removed (always bf16).

AOT (removal=false, barrier on qvalue): compiles through bwd; bwd all-reduce 28->25.
NOTE per-EXPERT [exp,1,mlp] breaks stock gmm_v2 bwd (_dlhs_scale_grad_by_rhs_scale repeats by
global group_sizes[256] vs 32 local experts); [1,1,mlp] shared sidesteps it.

+++ src/maxtext/kernels/megablox/ops.py
@@ -494,9 +494,12 @@ def _gmm_bwd(
   # TODO(tgale, enriqueps, apaske): Fuse this transposition into the tgmm.
   drhs = drhs.swapaxes(1, 2) if transpose_rhs else drhs
   if _orig_rhs_qarray is not None:
-    # Match the QArray pytree of the primal rhs; zero scale cotangent (forward-only fp8 weight-AG).
+    # The primal rhs entered as a QArray (moe_fp8_boundary_qag), so the VJP cotangent must match its
+    # pytree. Keep the qvalue-leaf cotangent (the weight gradient) in its NATURAL bf16/f32 dtype --
+    # NEVER cast it to e4m3 (that overflowed -> NaN and blew up the weight reduce-scatter). The scale
+    # is forward-only (STE on the quant), so its cotangent is a per-channel zero.
     drhs = qpl.QArray(
-        qvalue=drhs.astype(_orig_rhs_qarray.qvalue.dtype),
+        qvalue=drhs,
         scale=jnp.zeros_like(_orig_rhs_qarray.scale),
         zero_point=None,
         qtype=_orig_rhs_qarray.qtype,
